# prototype-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments": {
      "post": {
        "operationId": "createComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Comment.",
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "listComments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Comment records.",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{id}": {
      "get": {
        "operationId": "getComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Get a Comment by ID.",
        "tags": [
          "comments"
        ]
      },
      "patch": {
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Update a Comment.",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted."
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Delete a Comment.",
        "tags": [
          "comments"
        ]
      }
    },
    "/decisions": {
      "post": {
        "operationId": "createDecision",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Decision.",
        "tags": [
          "decisions"
        ]
      },
      "get": {
        "operationId": "listDecisions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Decision records.",
        "tags": [
          "decisions"
        ]
      }
    },
    "/decisions/{id}": {
      "get": {
        "operationId": "getDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Decision ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Decision not found."
          }
        },
        "summary": "Get a Decision by ID.",
        "tags": [
          "decisions"
        ]
      },
      "patch": {
        "operationId": "updateDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Decision ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Decision not found."
          }
        },
        "summary": "Update a Decision.",
        "tags": [
          "decisions"
        ]
      },
      "delete": {
        "operationId": "deleteDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Decision ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Decision deleted."
          },
          "404": {
            "description": "Decision not found."
          }
        },
        "summary": "Delete a Decision.",
        "tags": [
          "decisions"
        ]
      }
    },
    "/delegations": {
      "post": {
        "operationId": "createDelegation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDelegationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Delegation.",
        "tags": [
          "delegations"
        ]
      },
      "get": {
        "operationId": "listDelegations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DelegationResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Delegation records.",
        "tags": [
          "delegations"
        ]
      }
    },
    "/delegations/{id}": {
      "get": {
        "operationId": "getDelegation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Delegation ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Delegation not found."
          }
        },
        "summary": "Get a Delegation by ID.",
        "tags": [
          "delegations"
        ]
      },
      "patch": {
        "operationId": "updateDelegation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Delegation ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDelegationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Delegation not found."
          }
        },
        "summary": "Update a Delegation.",
        "tags": [
          "delegations"
        ]
      },
      "delete": {
        "operationId": "deleteDelegation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Delegation ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delegation deleted."
          },
          "404": {
            "description": "Delegation not found."
          }
        },
        "summary": "Delete a Delegation.",
        "tags": [
          "delegations"
        ]
      }
    },
...
  },
  "info": {
    "title": "IFS Prototype API",
    "description": "OpenAPI contract for the IFS prototype backend.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Comment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Comment\" for Comment entities.",
            "enum": [
              "Comment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this comment was last updated.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Comment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Comment\" for Comment entities.",
            "enum": [
              "Comment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this comment was last updated.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Comment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Comment\" for Comment entities.",
            "enum": [
              "Comment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this comment was last updated.",
            "format": "date-time"
          }
        }
      },
      "CreateDecisionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this decision."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Decision."
          },
          "entityType": {
            "type": "string",
            "description": "Entity category for this object, normally Decision."
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "subject": {
            "type": "string",
            "description": "Stable IFS reference to the object, proposal, rule, resource, or question being decided. Kept as a reference string because a decision subject can point to heterogeneous resources.",
            "format": "ifs-ref"
          },
          "state": {
            "enum": [
              "open",
              "closed",
              "rejected",
              "cancelled"
            ],
            "type": "string",
            "description": "Current lifecycle state of the decision."
          },
          "previousRevisionId": {
            "type": "string",
            "description": "Id of previous revision of this decision, If this decision is a modified version of another decision."
          },
          "previousRevision": {
            "description": "Current outcome of the decision, if the decision has produced an outcome at least once.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DecisionResponseDto"
              }
            ]
          },
          "eligibleMembers": {
            "description": "Member entities eligible to participate in this decision.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponseDto"
            }
          },
          "affectedMembers": {
            "description": "Member entities affected by this decision under the IFS scope rule.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponseDto"
            }
          },
          "votes": {
            "description": "Vote entities cast directly by members or by delegates acting on their behalf.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteResponseDto"
            }
          },
          "rules": {
            "description": "Rule entities that govern eligibility, delegation, quorum, thresholds, timing, or outcome calculation for this decision.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleResponseDto"
            }
          },
          "eligibleMemberCount": {
            "type": "number",
            "description": "Snapshot count of members eligible to participate.",
            "minimum": 0
          },
          "affectedMemberCount": {
            "type": "number",
            "description": "Snapshot count of members affected by this decision.",
            "minimum": 0
          },
          "voteCount": {
            "type": "number",
            "description": "Snapshot count of votes currently recorded for this decision.",
            "minimum": 0
          },
          "delegateVoteCount": {
            "type": "number",
            "description": "Snapshot count of votes cast by delegates on behalf of eligible members.",
            "minimum": 0
          },
          "openedAt": {
            "type": "string",
            "description": "Timestamp when this decision opened for participation.",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "description": "Optional timestamp when this decision closed.",
            "format": "date-time"
          },
          "cancelledAt": {
            "type": "string",
            "description": "Optional timestamp when this decision was cancelled.",
            "format": "date-time"
          },
          "verifiedAt": {
            "type": "string",
            "description": "Optional timestamp when participation, vote data, or outcome was verified.",
            "format": "date-time"
          },
          "approvedAt": {
            "type": "string",
            "description": "Optional timestamp when this decision produced an approved outcome.",
            "format": "date-time"
          },
          "rejectedAt": {
            "type": "string",
            "description": "Optional timestamp when this decision was rejected or produced a rejected outcome.",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this decision record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this decision record was last updated.",
            "format": "date-time"
          },
          "decidedAt": {
            "type": "string",
            "description": "Timestamp when this decision record was decided.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "subject",
          "state",
          "eligibleMembers",
          "affectedMembers",
          "votes",
          "eligibleMemberCount",
          "affectedMemberCount",
          "voteCount",
          "openedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateDecisionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this decision."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Decision."
          },
          "entityType": {
            "type": "string",
            "description": "Entity category for this object, normally Decision."
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "subject": {
            "type": "string",
            "description": "Stable IFS reference to the object, proposal, rule, resource, or question being decided. Kept as a reference string because a decision subject can point to heterogeneous resources.",
            "format": "ifs-ref"
          },
          "state": {
            "enum": [
              "open",
              "closed",
              "rejected",
              "cancelled"
            ],
            "type": "string",
            "description": "Current lifecycle state of the decision."
          },
          "previousRevisionId": {
            "type": "string",
            "description": "Id of previous revision of this decision, If this decision is a modified version of another decision."
          },
          "previousRevision": {
            "description": "Current outcome of the decision, if the decision has produced an outcome at least once.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DecisionResponseDto"
              }
            ]
          },
          "eligibleMembers": {
            "description": "Member entities eligible to participate in this decision.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponseDto"
            }
          },
          "affectedMembers": {
            "description": "Member entities affected by this decision under the IFS scope rule.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponseDto"
            }
          },
          "votes": {
            "description": "Vote entities cast directly by members or by delegates acting on their behalf.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteResponseDto"
            }
          },
          "rules": {
            "description": "Rule entities that govern eligibility, delegation, quorum, thresholds, timing, or outcome calculation for this decision.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleResponseDto"
            }
          },
          "eligibleMemberCount": {
            "type": "number",
            "description": "Snapshot count of members eligible to participate.",
            "minimum": 0
          },
          "affectedMemberCount": {
            "type": "number",
            "description": "Snapshot count of members affected by this decision.",
            "minimum": 0
          },
          "voteCount": {
            "type": "number",
            "description": "Snapshot count of votes currently recorded for this decision.",
            "minimum": 0
          },
          "delegateVoteCount": {
            "type": "number",
            "description": "Snapshot count of votes cast by delegates on behalf of eligible members.",
            "minimum": 0
          },
          "openedAt": {
            "type": "string",
            "description": "Timestamp when this decision opened for participation.",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "description": "Optional timestamp when this decision closed.",
            "format": "date-time"
          },
          "cancelledAt": {
            "type": "string",
            "description": "Optional timestamp when this decision was cancelled.",
            "format": "date-time"
          },
          "verifiedAt": {
            "type": "string",
            "description": "Optional timestamp when participation, vote data, or outcome was verified.",
            "format": "date-time"
          },
          "approvedAt": {
            "type": "string",
            "description": "Optional timestamp when this decision produced an approved outcome.",
            "format": "date-time"
          },
          "rejectedAt": {
            "type": "string",
            "description": "Optional timestamp when this decision was rejected or produced a rejected outcome.",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this decision record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this decision record was last updated.",
            "format": "date-time"
          },
          "decidedAt": {
            "type": "string",
            "description": "Timestamp when this decision record was decided.",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}