# prototype-api/src/comment/comment.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateCommentDto } from './dto/create-comment.dto.js';
import { CommentResponseDto } from './dto/comment-response.dto.js';
import { UpdateCommentDto } from './dto/update-comment.dto.js';
import { CommentService } from './comment.service.js';

@ApiTags('comments')
@Controller('comments')
export class CommentController {
  constructor(private readonly commentService: CommentService) {}

  @Post()
  @ApiOperation({ operationId: 'createComment', summary: 'Create a Comment.' })
  @ApiBody({ type: CreateCommentDto })
  @ApiCreatedResponse({ type: CommentResponseDto })
  create(@Body() data: CreateCommentDto): Promise<CommentResponseDto> {
    return this.commentService.create(data);
  }

  @Get()
  @ApiOperation({
    operationId: 'listComments',
    summary: 'List Comment records.',
  })
  @ApiOkResponse({ type: CommentResponseDto, isArray: true })
  findAll(): Promise<CommentResponseDto[]> {
    return this.commentService.findAll();
  }

  @Get(':id')
  @ApiOperation({ operationId: 'getComment', summary: 'Get a Comment by ID.' })
  @ApiParam({ name: 'id', type: String, description: 'Comment ID.' })
  @ApiOkResponse({ type: CommentResponseDto })
  @ApiNotFoundResponse({ description: 'Comment not found.' })
  async findOne(@Param('id') id: string): Promise<CommentResponseDto> {
    const result = await this.commentService.findOne(id);
    if (!result) throw new NotFoundException('Comment not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({ operationId: 'updateComment', summary: 'Update a Comment.' })
  @ApiParam({ name: 'id', type: String, description: 'Comment ID.' })
  @ApiBody({ type: UpdateCommentDto })
  @ApiOkResponse({ type: CommentResponseDto })
  @ApiNotFoundResponse({ description: 'Comment not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateCommentDto,
  ): Promise<CommentResponseDto> {
    return this.commentService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ operationId: 'deleteComment', summary: 'Delete a Comment.' })
  @ApiParam({ name: 'id', type: String, description: 'Comment ID.' })
  @ApiNoContentResponse({ description: 'Comment deleted.' })
  @ApiNotFoundResponse({ description: 'Comment not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.commentService.remove(id);
  }
}