# prototype-api/src/decision/decision.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateDecisionDto } from './dto/create-decision.dto.js';
import { DecisionResponseDto } from './dto/decision-response.dto.js';
import { UpdateDecisionDto } from './dto/update-decision.dto.js';
import { DecisionService } from './decision.service.js';

@ApiTags('decisions')
@Controller('decisions')
export class DecisionController {
  constructor(private readonly decisionService: DecisionService) {}

  @Post()
  @ApiOperation({
    operationId: 'createDecision',
    summary: 'Create a Decision.',
  })
  @ApiBody({ type: CreateDecisionDto })
  @ApiCreatedResponse({ type: DecisionResponseDto })
  create(@Body() data: CreateDecisionDto): Promise<DecisionResponseDto> {
    return this.decisionService.create(data);
  }

  @Get()
  @ApiOperation({
    operationId: 'listDecisions',
    summary: 'List Decision records.',
  })
  @ApiOkResponse({ type: DecisionResponseDto, isArray: true })
  findAll(): Promise<DecisionResponseDto[]> {
    return this.decisionService.findAll();
  }

  @Get(':id')
  @ApiOperation({
    operationId: 'getDecision',
    summary: 'Get a Decision by ID.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Decision ID.' })
  @ApiOkResponse({ type: DecisionResponseDto })
  @ApiNotFoundResponse({ description: 'Decision not found.' })
  async findOne(@Param('id') id: string): Promise<DecisionResponseDto> {
    const result = await this.decisionService.findOne(id);
    if (!result) throw new NotFoundException('Decision not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({
    operationId: 'updateDecision',
    summary: 'Update a Decision.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Decision ID.' })
  @ApiBody({ type: UpdateDecisionDto })
  @ApiOkResponse({ type: DecisionResponseDto })
  @ApiNotFoundResponse({ description: 'Decision not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateDecisionDto,
  ): Promise<DecisionResponseDto> {
    return this.decisionService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({
    operationId: 'deleteDecision',
    summary: 'Delete a Decision.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Decision ID.' })
  @ApiNoContentResponse({ description: 'Decision deleted.' })
  @ApiNotFoundResponse({ description: 'Decision not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.decisionService.remove(id);
  }
}