# prototype-api/src/delegation/delegation.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateDelegationDto } from './dto/create-delegation.dto.js';
import { DelegationResponseDto } from './dto/delegation-response.dto.js';
import { UpdateDelegationDto } from './dto/update-delegation.dto.js';
import { DelegationService } from './delegation.service.js';

@ApiTags('delegations')
@Controller('delegations')
export class DelegationController {
  constructor(private readonly delegationService: DelegationService) {}

  @Post()
  @ApiOperation({
    operationId: 'createDelegation',
    summary: 'Create a Delegation.',
  })
  @ApiBody({ type: CreateDelegationDto })
  @ApiCreatedResponse({ type: DelegationResponseDto })
  create(@Body() data: CreateDelegationDto): Promise<DelegationResponseDto> {
    return this.delegationService.create(data);
  }

  @Get()
  @ApiOperation({
    operationId: 'listDelegations',
    summary: 'List Delegation records.',
  })
  @ApiOkResponse({ type: DelegationResponseDto, isArray: true })
  findAll(): Promise<DelegationResponseDto[]> {
    return this.delegationService.findAll();
  }

  @Get(':id')
  @ApiOperation({
    operationId: 'getDelegation',
    summary: 'Get a Delegation by ID.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Delegation ID.' })
  @ApiOkResponse({ type: DelegationResponseDto })
  @ApiNotFoundResponse({ description: 'Delegation not found.' })
  async findOne(@Param('id') id: string): Promise<DelegationResponseDto> {
    const result = await this.delegationService.findOne(id);
    if (!result) throw new NotFoundException('Delegation not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({
    operationId: 'updateDelegation',
    summary: 'Update a Delegation.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Delegation ID.' })
  @ApiBody({ type: UpdateDelegationDto })
  @ApiOkResponse({ type: DelegationResponseDto })
  @ApiNotFoundResponse({ description: 'Delegation not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateDelegationDto,
  ): Promise<DelegationResponseDto> {
    return this.delegationService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({
    operationId: 'deleteDelegation',
    summary: 'Delete a Delegation.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Delegation ID.' })
  @ApiNoContentResponse({ description: 'Delegation deleted.' })
  @ApiNotFoundResponse({ description: 'Delegation not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.delegationService.remove(id);
  }
}